+++ app/Jobs/ProcessRespondingSignRequest.php
@@ -11,6 +11,7 @@ use Illuminate\Foundation\Bus\Dispatchable;
 use Illuminate\Queue\InteractsWithQueue;
 use Illuminate\Queue\SerializesModels;
 use Illuminate\Support\Facades\Notification;
+use Illuminate\Support\Facades\Storage;
 
 class ProcessRespondingSignRequest implements ShouldQueue
 {
@@ -34,10 +35,11 @@ class ProcessRespondingSignRequest implements ShouldQueue
      */
     public function handle()
     {
-        $encrypted_signed_filename = hash('sha256', $this->data['requester']['email'] . $this->data['requester']['filename'] . $this->data['requester']['id'] . '-Signed');
+        $encrypted_filename = hash('sha256', $this->data['requester']['email'] . $this->data['requester']['filename'] . $this->data['requester']['id'] . '-Signed');
+        $encrypted_unsigned_filename = hash('sha256', $this->data['requester']['email'] . $this->data['requester']['filename'] . $this->data['requester']['id']);
         Notification::route('mail', $this->data['email'])->notify(new Signer($this->data));
         Notification::route('mail', $this->data['requester']['email'])->notify(new Requester($this->data));
-        dispatch(new ProcessDeleteFIle($encrypted_signed_filename))->delay(now()->addMinutes(30));
-
+        dispatch(new ProcessDeleteFIle($encrypted_filename))->delay(now()->addMinutes(15));
+        Storage::delete("requested_files/{$encrypted_unsigned_filename}");
     }
 }