+++ resources/js/signrequestapp/i18n/en/confirmSubmitModal.js
@@ -1,7 +1,7 @@
 export default {
   'en.confirmSubmitModal': {
     'title': 'ARE YOU SURE?',
-    'body': 'Your document will be emailed to your signer and you as a requester. Your document will be deleted in 12 hours from now.',
+    'body': 'Your document will be emailed to your signer and you as a requester. Your document will be deleted in 12 hours or after it signed.',
     'button': {
       'submit': 'Submit',
       'cancel': 'Cancel'