+++ resources/js/signrequestapp/i18n/id/confirmSubmitModal.js
@@ -1,7 +1,7 @@
 export default {
   'id.confirmSubmitModal': {
     'title': 'APAKAH ANDA YAKIN?',
-    'body': 'Dokumen anda akan dikirimkan ke alamat email yang anda masukan sebagai penandatangan dan ke email anda sebagai pemohon',
+    'body': 'Dokumen anda akan dikirimkan ke alamat email yang anda masukan sebagai penandatangan dan ke email anda sebagai pemohon. Dokumen anda akan dihapus dalam 12 jam dari sekarang atau setelah ditanda-tangani',
     'button': {
       'submit': 'Kirim',
       'cancel': 'Batal'