+++ resources/js/signrequestrespondapp/i18n/en/index.js
@@ -1,12 +1,14 @@
 import signaturePad from './signaturePad'
 import endModal from './endModal'
 import sharingSocial from './sharingSocial'
+import feedback from './feedback'
 
 
 export default {
 
   ...signaturePad,
   ...endModal,
-  ...sharingSocial
+  ...sharingSocial,
+  ...feedback
 
 }