+++ resources/js/signrequestrespondapp/i18n/id/index.js
@@ -1,11 +1,13 @@
 import signaturePad from './signaturePad'
 import endModal from './endModal'
 import sharingSocial from './sharingSocial'
+import feedback from './feedback'
 
 export default {
 
   ...signaturePad,
   ...endModal,
-  ...sharingSocial
+  ...sharingSocial,
+  ...feedback
 
 }